+++ simulator/messages/vessellogbookoutputmessage.h
@@ -19,7 +19,7 @@ public:
     }
 
     bool send(std::ostream &);
-    virtual int sendBinary(void *buffer, size_t maxlen);
+    virtual size_t sendBinary(void *buffer, size_t maxlen);
 
 private:
     std::ostream &loglike;
