+++ simulator/outputqueuemanager.cpp
@@ -22,7 +22,7 @@ public:
         return false;
     }
 
-    int sendBinary(void *buffer, size_t maxlen){
+    size_t sendBinary(void *buffer, size_t maxlen){
         UNUSED(buffer);
         UNUSED(maxlen);
